Week 2 Homework - LocationFragment done.
Description: I moved to code from MainActivity to LocationFragment. And navigate it with Jetpack Navigation library. I cleaned to MainActivity and refactor on LocationFragment.\nI used appcompat searchview. It has default options such as text color and font. I changed the color of text but I'm still trying to change its font.\nAnd yes I used places API to search places but I deleted it. The reason why I tried to use it because whenever I search a location for example Ankara, it gives me hint when I write it. That's super easy but it's like black box. I couldn't change it's default theme. So that's why I deleted it. But if I have time, I will try to integrate it too.